LayoutTests/editing/text-iterator/invalid-subrange-expected.txt

PASS internals.lengthFromRange(container, subrange) is 0
PASS internals.lengthFromRange(container, subrange) is 0
PASS successfullyParsed is true

TEST COMPLETE
